Switch head from AtomicInteger to AtomicLong (#8490)
* Bring the modulo operation inside the increment to prevent overflow errors

* Switch head to be an AtomicLong

* fix

+++ core/jvm/src/main/scala/zio/internal/metrics/ConcurrentMetricHooksPlatformSpecific.scala
@@ -125,7 +125,7 @@ private[zio] class ConcurrentMetricHooksPlatformSpecific extends ConcurrentMetri
     import key.keyType.{maxSize, maxAge, error, quantiles}
 
     val values = new AtomicReferenceArray[(Double, java.time.Instant)](maxSize)
-    val head   = new AtomicInteger(0)
+    val head   = new AtomicLong(0)
     val count  = new LongAdder
     val sum    = new DoubleAdder
     val min    = AtomicDouble.make(Double.MaxValue)
@@ -174,7 +174,7 @@ private[zio] class ConcurrentMetricHooksPlatformSpecific extends ConcurrentMetri
     // While Observing we cut off the first sample if we have already maxSize samples
     def observe(tuple: (Double, java.time.Instant)): Unit = {
       if (maxSize > 0) {
-        val target = head.incrementAndGet() % maxSize
+        val target = (head.incrementAndGet() % maxSize).toInt
         values.set(target, tuple)
       }
